+++ src/main/java/racingcar/Main.java
@@ -1,7 +1,9 @@
 package racingcar;
 
-import racingcar.ui.InputView;
-import racingcar.ui.ResultView;
+import racingcar.domain.RacingCar;
+import racingcar.domain.RandomNumber;
+import racingcar.view.InputView;
+import racingcar.view.ResultView;
 
 import java.util.List;
