+++ src/main/java/racingcar/domain/RandomNumber.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.domain;
 
-import racingcar.util.RandomGenerator;
+import racingcar.utils.RandomGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
