+++ src/test/java/racingcar/RacingCarTest.java
@@ -4,7 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.ui.ResultView;
+import racingcar.domain.Car;
+import racingcar.domain.RacingCar;
+import racingcar.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.Arrays;
