+++ src/test/java/racingcar/RandomNumberTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RandomNumber;
 
 import java.util.List;
